Fix-libdir
The previous patch for the below bug report used $ENV{LIBDIR} but this
wasn't enough to help with cross-compiling. We could use $ENV{SYSROOT}
but this results in a bogus RPATH entry. It is better to simply trust
the toolchain, which may have its own sysroot applied.

Cross-compiling Perl modules requires additional hackery but at least
that hackery will work now. :)

Bug: https://bugs.gentoo.org/416339